+++ app/(auth)/not-found.tsx
@@ -1,0 +1,10 @@
+import Link from "next/link";
+export default function NotFound() {
+    return (
+        <div>
+        <h1>404 - Page Not Found</h1>
+        <p>Sorry, the page you are looking for does not exist.</p>
+        <Link href="/">Go back to Home</Link>
+        </div>
+    );
+}
